Change long long to int

+++ SamsungSWEA_3431_ManagingJunHwan'sExercise/SamsungSWEA_3431_ManagingJunHwan'sExercise/SamsungSWEA_3431_ManagingJunHwan'sExercise.cpp
@@ -10,10 +10,10 @@ int main(int argc, char** argv)
 	cin >> T;
 	for (test_case = 1; test_case <= T; ++test_case)
 	{
-		long long L, U, X;
-		scanf("%lld %lld %lld", &L, &U, &X); //L<X<U
+		int L, U, X;
+		scanf("%d %d %d", &L, &U, &X); //L<X<U
 		if (X <= L) {
-			printf("#%d %lld\n", test_case, L - X);
+			printf("#%d %d\n", test_case, L - X);
 		}
 		else if (X < U) {
 			printf("#%d 0\n", test_case);
